Fix: 파일 위치 수정 및 중복 파일 제거
