+++ src/main/java/com/hotel/flint/reserve/room/dto/RoomReservedDto.java
@@ -2,7 +2,7 @@ package com.hotel.flint.reserve.room.dto;
 
 import com.hotel.flint.common.enumdir.Option;
 import com.hotel.flint.reserve.room.domain.ReservedRoom;
-import com.hotel.flint.reserve.room.domain.RoomDetails;
+import com.hotel.flint.room.domain.RoomDetails;
 import com.hotel.flint.reserve.room.domain.RoomReservation;
 import com.hotel.flint.user.member.domain.Member;
 import lombok.AllArgsConstructor;
