+++ src/main/java/com/hotel/flint/room/repository/RoomDetailsRepository.java
@@ -1,6 +1,6 @@
-package com.hotel.flint.reserve.room.repository;
+package com.hotel.flint.room.repository;
 
-import com.hotel.flint.reserve.room.domain.RoomDetails;
+import com.hotel.flint.room.domain.RoomDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomDetailsRepository extends JpaRepository<RoomDetails, Long> {
