+++ src/main/java/com/hotel/flint/room/repository/RoomInfoRepository.java
@@ -1,6 +1,6 @@
-package com.hotel.flint.reserve.room.repository;
+package com.hotel.flint.room.repository;
 
-import com.hotel.flint.reserve.room.domain.RoomInfo;
+import com.hotel.flint.room.domain.RoomInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomInfoRepository extends JpaRepository<RoomInfo, Long> {
