+++ src/main/java/com/hotel/flint/user/employee/dto/InfoRoomResDto.java
@@ -1,8 +1,5 @@
 package com.hotel.flint.user.employee.dto;
 
-import com.hotel.flint.reserve.room.domain.RoomDetails;
-import com.hotel.flint.reserve.room.domain.RoomReservation;
-import com.hotel.flint.user.member.domain.Member;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
