+++ src/main/resources/application.yml
@@ -9,7 +9,7 @@ spring:
     database-platform: org.hibernate.dialect.MariaDBDialect
     generate-ddl: true
     hibernate:
-      ddl-auto: update
+      ddl-auto: create
     show_sql: true
 # 사원증 만들때 사용하기 위해 설정.
 fixed:
